Update mods.jsonc for Level Extender
updated Level Extender, added current UniqueID, added GitHub link to current repo. 

I'm the current lead maintainer on the mod, Unidarkshin is still the author and has the ability to contribute to the repo. Not sure if my username needs to be reflected in here or not, left it out for now